Move webpack config and package json to assets folder. Add Make file. (#296)
* Move webpack config and package json to assets folder. Add Make file

* remove manifest.json

* node_modules is already ignored below

* Add make to travis script file
