+++ .gitignore
@@ -8,7 +8,6 @@ instance/testing.py
 .webassets-cache
 baseframe-packed.js
 baseframe-packed.css
-boxoffice/static/js/node_modules/
 boxoffice/static/js/npm-debug.log
 .vscode
 .coverage
@@ -16,3 +15,5 @@ boxoffice/static/gen/*
 *.sql
 *.sql.gz
 *.sql.bz2
+node_modules
+boxoffice/static/build
