+++ .travis.yml
@@ -14,6 +14,7 @@ install:
   - pip install -r requirements.txt
   - pip install https://github.com/hasgeek/coaster/archive/master.zip
   - npm install casperjs
+  - make
 env:
   - FLASK_ENV=testing
 script:
