+++ Makefile
@@ -1,0 +1,2 @@
+all:
+	cd boxoffice/assets; make
