+++ boxoffice/assets/Makefile
@@ -1,0 +1,8 @@
+all: assets
+
+assets:
+	npm install
+	npm run build
+
+assetsonly:
+	npm run build
