--- boxoffice/static/package.json
@@ -1,35 +1,0 @@
-{
-  "scripts": {
-    "dev": "NODE_ENV=development node_modules/.bin/webpack --progress --profile --colors",
-    "build": "NODE_ENV=production node_modules/.bin/webpack --progress --profile --colors"
-  },
-  "dependencies": {
-    "babel-core": "^6.26.0",
-    "babel-loader": "^7.1.2",
-    "babel-preset-es2015": "^6.24.1",
-    "backbone": "~1.3.3",
-    "c3": "~0.4.11",
-    "clipboard": "^1.6.0",
-    "css-loader": "^0.14.4",
-    "d3": "~3.5.17",
-    "extract-text-webpack-plugin": "^3.0.2",
-    "html-webpack-plugin": "^2.30.1",
-    "image-webpack-loader": "^1.4.0",
-    "jquery": "^3.2.1",
-    "moment": "^2.18.1",
-    "nprogress": "^0.2.0",
-    "ractive": "^0.7.3",
-    "ractive-transitions-fly": "^0.3.0",
-    "rome": "^2.1.22",
-    "script-loader": "^0.6.1",
-    "select2": "~4.0.3",
-    "style-loader": "^0.12.3",
-    "underscore": "~1.8.3",
-    "validate-js": "^2.0.1",
-    "webpack": "^3.8.1",
-    "webpack-dev-server": "^2.9.4"
-  },
-  "devDependencies": {
-    "extract-text-webpack-plugin-webpack-2": "^1.0.1"
-  }
-}
